requirements.txt
boto==2.49.0
boto3==1.12.16
botocore==1.15.16
certifi==2019.11.28
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.4
djangorestframework==3.10.2
docutils==0.15.2
gensim==3.8.1
gunicorn==20.0.4
idna==2.9
jmespath==0.9.5
joblib==0.14.1
nltk==3.4.5
numpy==1.20.1
pandas==1.3.1
psycopg2-binary==2.9.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
scikit-learn==0.24.2
scipy==1.7.1
six==1.14.0
smart-open==1.9.0
sqlparse==0.3.1
threadpoolctl==2.2.0
urllib3==1.25.8
